fix: dockerfile
### Problem
In current scenario we're taking the `.env` variables from the production env file but that file was not being pushed to central code repository system i.e., `GitHub` and therefore build thus formed by ci/cd pipeline was not correct. So, to fix this issue we're taking the env variables at run-time.